+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 ### Fixed:
 - "Show Setting" after running `motionbuilder.setupCodeCompletion` not always showing the correct setting
 
+### Removed:
+- Deprecated settings: `motionbuilder.debug.port` & `motionbuilder.debug.autoPort`
+
 
 ## [2025.0.2] - 2024-05-05